# docs/public/reference/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_version_pipelines__pipeline_id__versions_upload_post": {
        "properties": {
          "config_file": {
            "format": "binary",
            "title": "Config File",
            "type": "string"
          },
          "registered_by": {
            "title": "Registered By",
            "type": "string"
          },
          "scripts": {
            "format": "binary",
            "title": "Scripts",
            "type": "string"
          }
        },
        "required": [
          "config_file",
          "scripts",
          "registered_by"
        ],
        "title": "Body_upload_version_pipelines__pipeline_id__versions_upload_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RegisterPipeline": {
        "properties": {
          "department_id": {
            "title": "Department Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "registered_by": {
            "title": "Registered By",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "tenant_id",
          "name",
          "department_id",
          "registered_by"
        ],
        "title": "RegisterPipeline",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Munitas control plane",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/pipelines": {
      "get": {
        "description": "Every pipeline in the caller's own tenant, with the total for real\npage controls.\n\nPreviously took a caller-supplied `tenant_id` with no session at all --\nany tenant's full pipeline list, names, departments and who registered\nthem, readable by guessing the tenant id. Scoped to the session's own\ntenant now, the same fix every read endpoint in `read_models.py`\nalready has for this exact shape of gap.\n\nAlso previously had no `limit`/`offset` at all, unlike every other list\nendpoint -- missed in the console-wide pagination pass this was found\nduring, since nothing had ever flagged it as unbounded before an\norganisation actually had enough pipelines for that to matter. Same\n`{pipelines, shown, total, limit, offset}` envelope as the rest.",
        "operationId": "list_pipelines_pipelines_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response List Pipelines Pipelines Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Pipelines",
        "tags": [
          "pipelines"
        ]
      }
    },
    "/pipelines/register": {
      "post": {
        "description": "Register a pipeline. No version yet, so nothing about it can run\nuntil one is sealed, the same rule an agent's registration already\nfollows.",
        "operationId": "register_pipelines_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterPipeline"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Register Pipelines Register Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register",
        "tags": [
          "pipelines"
        ]
      }
    },
    "/pipelines/{pipeline_id}": {
      "get": {
        "description": "One pipeline and its versions, scoped to the caller's own tenant.\n\nPreviously took no session at all: any pipeline's config and version\nhistory, by anyone who knew or guessed its id, cross-tenant.",
        "operationId": "read_pipeline_pipelines__pipeline_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "pipeline_id",
            "required": true,
            "schema": {
              "title": "Pipeline Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Read Pipeline Pipelines  Pipeline Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Pipeline",
        "tags": [
          "pipelines"
        ]
      }
    },
    "/pipelines/{pipeline_id}/versions/upload": {
      "post": {
        "description": "Upload a version: one YAML file naming the steps, one zip holding every\nscript those steps reference. Sealed on arrival, the same as an agent\nversion: this platform never lets a version be edited after the fact,\nonly superseded by the next one.",
        "operationId": "upload_version_pipelines__pipeline_id__versions_upload_post",
        "parameters": [
          {
            "in": "path",
            "name": "pipeline_id",
            "required": true,
            "schema": {
              "title": "Pipeline Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_version_pipelines__pipeline_id__versions_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Upload Version Pipelines  Pipeline Id  Versions Upload Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Version",
        "tags": [
          "pipelines"
        ]
      }
    },
    "/pipelines/{pipeline_id}/versions/{version_id}/code": {
      "get": {
        "description": "Stream a version's scripts.zip back, for the worker only.\n\nGated the same way agent_upload.py's own download_code is: a shared\nsecret, not a directory identity, because the caller is the worker\nprocess itself. Never reachable from agentnet.",
        "operationId": "download_code_pipelines__pipeline_id__versions__version_id__code_get",
        "parameters": [
          {
            "in": "path",
            "name": "pipeline_id",
            "required": true,
            "schema": {
              "title": "Pipeline Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-worker-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Worker-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Code",
        "tags": [
          "pipelines"
        ]
      }
    },
...
  }
}